Tests for txn manager services begin, commit, abort and building txn dependency list.
Documentatin for txn manager.


git-svn-id: https://svn.apache.org/repos/asf/directory/apacheds/branches/apacheds-txns@1195429 13f79535-47bb-0310-9956-ffa450edef68

+++ core/src/main/java/org/apache/directory/server/core/txn/TxnManagerFactory.java
@@ -58,7 +58,7 @@ public class TxnManagerFactory
      */
     @SuppressWarnings("unchecked")
     public static <ID> void init( Comparator<ID> idComparator, Serializer idSerializer, String logFolderPath,
-        int logBufferSize, int logFileSize ) throws IOException
+        int logBufferSize, long logFileSize ) throws IOException
     {
         Log log = new DefaultLog();
